My first website about corns
I'm just starting in programming and making websites so if there are some issues with the website write a report, I'll be glad to fix it. And also DO NOT open it from phone because there's gonna be a mess, you won't be able to read information, this website IS NOT adapted by phones if someone know how to make the website adaptable please leave the comment, I should also mention that this is my first time of using GitHub so I don't know many features there.